+++ src/library_nodefs.js
@@ -168,9 +168,10 @@ addToLibrary({
           // update the common node structure mode as well
           node.mode = attr.mode;
         }
-        if (attr.timestamp !== undefined) {
-          var date = new Date(attr.timestamp);
-          utimes(arg, date, date);
+        if (attr.atime || attr.mtime) {
+          var atime = attr.atime && new Date(attr.atime);
+          var mtime = attr.mtime && new Date(attr.mtime);
+          fs.utimesSync(path, atime, mtime);
         }
         if (attr.size !== undefined) {
           truncate(arg, attr.size);